agents-cloud-run/pyproject.toml
[project]
name = "adk-cloud-run"
version = "0.1.0"
description = "ADK Agent for Google Cloud Run"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.16",
    "starlette>=0.46.1",
    "uvicorn>=0.34.0",
    "click>=8.1.8",
    "google-adk>=1.7.0",
    "python-dotenv>=1.1.0",
    "asyncpg>=0.30.0",
    "google-cloud-alloydb-connector[asyncpg]>=1.9.0",
    "litellm>=1.74.8",
]
